Initial hacky implementation
Proof-of-concept showing that this can actually work; find the audio
outputs that the Windows volume mixer displays (sndvol) and set their
volumes based on some defined configuration.

This implementation simply does an O(NM) search for suffix substrings
across all audio sessions and all applications, matching against their
FullProcessImageName. This means that the process needs to have
permission for PROCESS_QUERY_LIMITED_INFORMATION, but that doesn't
appear to require any additional configuration. I'm not sure what
permissions are necessary to actually change the volume, but this
program does not need to be run as Administrator.