Update clang-format from default
LLVM style with conditionals allowed on a single line. My actual
personal rule is that braces are omitted iff the conditional is on a
single line. This allows for concise early exits (which are always
encouraged) hopefully without the danger of conditionals lacking
braces. However, I don't think clang-format supports linting that.