Set volume configuration with a TOML file
This didn't quite go according to plan in a few ways, but it'll do the
job for now. Having not used TOML with C++ before I don't have a
library preference, so I just went with `toml11`. Easy enough to use,
but I don't think it supports `std::wstring` so we're changing to UTF-8.
I wanted to that anyway at some point; we're doing it now.

For my usage it's sufficient to define a number of 'profiles', where
each profile is simply a list of suffix/volume pairs. When a profile is
active, the application matches each executable path to the suffixes and
sets the appropriate volume for each match. I therefore expected to be
able to write something like
```
[default]
"*" = 0.3
"\\chrome.exe" = 0.5
```
to define a profile `default` that set the volume of the device to 0.3
and the volume of Chrome to 0.5 (of that 0.3). However, it's possible
that two suffixes match the same executable, and therefore some
prioritisation is needed. This could be used, for example, to assign a
volume for all instances of `foo.exe` except for the one in `/bar/`,
which sounds like a useful feature. The easiest way to set priorities is
to just use the key ordering, but since TOML tables are generic
hash-tables, they don't care about order.

Since overlap is inevitable with the format above, and leaving the order
up to the implementation makes me a little uneasy (not least because
with hash randomization the order would become nondeterministic), we're
stuck with the more verbose format given in the example. It's more
flexible and is fine, but I'm a little disappointed to not have the
simplicity of the original design.