Add support for setting system sounds volume
Now that we want to set both the device volume and the system volume,
there are two volume controls that aren't actually suffixes. I think
using "*" to mean the device does kind of make sense, because changing
the device volume does affect the volume of everything, it clearly
isn't extensible.

Anything extensible that could be used to consistently refer to the
device, system sounds, or other things that aren't `exe`s, must not be
a substring of any path. Otherwise, they would prevent certain `exe`s
from being matched. Looking through the list of reserved characters
shows that `:` is disallowed, and Ruby sets a precedent for prefixing
symbols with `:` so we go with that. `:device` for the device volume,
and `:system` for the system sounds (beeps and so on).

Of course ``C:\`` contains a `:` so it's not quite true that paths can't
contain one, but in a fully qualified path a `:` will always be followed
by a ``\`` and therefore we're safe to use alphanumeric characters after
the `:`. Relative paths may lack the first ``\``, but these will not be
returned by `QueryFullProcessImageName`.

This strategy has no hope of working on Linux, where I believe paths can
contain anything except nulls, but it's unlikely paths will contain a
`:` and we could probably use some actual file system path in those
cases. I haven't thought too hard about a Linux implementation yet.