Add back the function try block around main
The C++/WinRT documentation recommends against this because without it
the exception can be caught by the environment and displayed in a GUI
to the user, and the developer can catch it in a breakpoint if they're
running the program under a debugger.

I'm unconvinced; yes a GUI does pop up, but it omits the crucial
information of _what_ the exception actually is. None of the advertised
error report or minidump appears, just a prompt to abort or open in a
debugger, which doesn't actually do anything if the program wasn't
being run in a debugger in the first place. In that case, one must now
run the application again under a debugger in the exact same conditions
to replicate the crash. Yes if the program was being run in a debugger
originally then being able to break at the exception is useful, but
this particular program is simple enough that just knowing the
exception message is enough to debug the issue, saving frustratingly
running the program twice.

Worse, in release builds no such message appears and the program simply
crashes completely with an arcane exit code which even as a developer
is often a nightmare to decode.