Do not create app data config directory
If we have to create the directory then there can't be a config file
inside it, so the only point would be to set up the environment for the
user, which is the installer's job. Arguably the installer shouldn't
even do it in this case because there's no sensible default config file,
and the user can choose to not use that location at all by always
passing `--config`. We should not create a directory that the user does
not want.