Check volume levels are in a valid range
Let `VolumeControl` establish an invariant that the volume is within a
valid range. Checking at the point of parsing does not prevent
`VolumeControl`s being created with volumes outside that range by other
means, and checking at the call-site is tedious and error-prone.
If the invariant cannot be established then the caller can still
provide improvements on the error message like they would be able to do
with inline validation at parse-time.

This possibly unnecessarily restricts things on non-Windows platforms
where it might be possible to set the volume above 100%, but we
currently don't support any of those, and I believe those platforms
usually advise against setting the volume that high.