main.cpp -> executable.cpp
This is to disambiguate from the service executable, that will also
have a `main` function. Though the executable itself will probably be
named `declvol` in the future, the source file has not been named that
so as to not conflict with the implementation file of a possible future
common library header, which would conventionally be named `declvol.h`.