Add skeleton service implementation
The idea is to use a Windows service to listen to start process events,
then set the volume of the newly launched process as usual. This is
unlikely to be entirely straightforward due to audio not being ready
at process startup, and I'm not even sure if you can receive such
events, but it'll be very convenient if this works.

The state machine is kind of overengineered, it's not as clear as I'd
hoped it would be. I still think it's easier to reason about than
embedding it implicitly in `run_service`, however.