Get notified on process start using Windows MI
Not to be confused with WMI, the Windows Management Infrastructure
system is the latest version of the Windows Management Instrumentation
(WMI) system that is compatible but apparently easier to use.

This does at least seem to work, but I'm unsatisfied with the
security implications and whether it'll even work on Windows 10 Home
given that the `winrm` service must be enabled.