Implement destructor in MIBase
Rather than requiring the user to write the implementation of the
destructor in the derived class, move the destructor implementations
to an NTTP `Deleter`. Since the deleter implementations are very
short, this is straightforward for users and allows them to use the
rule of zero. Previously, the user-declared destructor in the
derived class would prevent the other special member functions from
being implicitly declared, requiring the user to explicitly default
them.