Replace call to OpenProcess with em::open_process
Because `open_process` throws an exception on error, rather than letting
the caller check the return value, keeping the same error handling
here would be complicated. However, since the system process has already
been excluded, there shouldn't be any issues with opening the process.

Additionally, while the previous code was continuing with the next
session if the process could not be opened, it was not continuing if
changing the session volume threw an exception. Be consistent and let
both escape to the caller.