Replace IMMDevice overload of get_audio_sessions
My reading of `IMMDevice::Activate` was that it would return a
reference to the given interface on the device, and that future calls
to `Activate` would return references to the same interface and in
particular the same _implementation_ of that interface. This is wrong,
multiple calls to `Activate` the same interface will return different
implementations.

This fact can be the source of some quite confusing bugs, so the control
of the `IAudioSessionManager2` must be left completely to the user.
In particular, library functions must not obtain a pointer to a manager
themselves, they must accept one from the user.