Replace service implementation with --wait flag
Well this has been a journey. Despite the flag that can be set in the
service manager, services cannot interact with the desktop environment
in any way, including when run as the user themselves. A service
implementation of this program therefore cannot work. [1] I probably
should have known that.

Instead, we give up on autostart and discard everything to get a
_much_ simpler implementation. Now the user simply runs the same
executable, but passes a `--wait` flag to tell the program to stay open
and listen for audio session creation events. Trying to get a nice
shutdown on CTRL-C on Windows also seems to be very hard, so we just
wait for the user to press enter.

I thought about collapsing the commit history down, but nobody's
reviewing this as a patch, and I think there's some interesting stuff
there if people do want to take a look. Including me, because the MI
and service code might be useful in the future for something else.

[1]: This isn't _quite_ true, as the service was still able to see the
     system audio session, and therefore could probably manipulate that.