Change profile of waiter through setter over gRPC
Because the waiter process (called with `--wait`) sets its configuration
on startup, a setter started after a waiter can use a different
profile to the waiter, and the two can disagree. The setter can even
use a different configuration file. Additionally, the user can change
the contents of the configuration file while the waiter is running, and
the waiter will not see those changes until it is restarted.

Although restarting the waiter after the configuration file is changed
might be acceptable, the waiter and setter disagreeing on the
configuration is not. When a setter is run, the waiter must change its
current profile to match.

There are numerous ways of doing this, and the method chosen here is
for the setter to notify the waiter over gRPC. With the size of the
dependencies this feels a little like hitting a tack with a
sledgehammer, but the actual code is straightforward and consists
mostly of boilerplate. One of the reasons to use gRPC is just that I'm
already familiar with it, and though I'm no expert, I've more experience
than with e.g. Cap'n Proto. I considered trying ZeroMQ, but again I
have no experience with it and message serialization would still have
to be dealt with. Its main attraction was using IPC rather than a
networking protocol, because that's really all this needs, but that
isn't currently supported on Windows. Using shared memory directly
with something like Boost.Interprocess would probably be better for the
user, but the synchronization and actual implementation would be much
more complicated. I think an IPC approach is worth pursuing, just not
right now.

The `tools/build_protos.sh` script (run in WSL) should be used, with
appropriate modification, to regenerate the Protobuf implementation
files whenever the Protobufs are changed. While in theory the
implementation could be generated by CMake, I prefer to have Protobuf
files checked into source control anyway, and honestly I just don't want
to write that build system code right now.