Fix consistency when two waiters try to start
The main problem was that waiters are actually also setters, but they
weren't sending notifications to the waiter after setting the volume,
which causes an inconsistent state when the existing waiter is using a
different profile to the new (failing to wait) waiter.

This commit checks that an existing waiter is running by just trying to
start the service; if another service is bound to that port, then it
will fail. I would call this the 'correct' way of doing it, where we
treat it like taking a lock on a mutex. In this case, it would actually
be better to do it the 'incorrect' way.

With the 'correct' way, if we succeed in binding then we're the sole
waiter, otherwise there may or may not be another waiter running.
No error code is returned to contain this error, instead we get a log
message from gRPC.

For a developer the message is great; in debug mode it has the full path
to the file, and it has multiple levels of explanation wrapped up in
some serialized JSON. For a user, it's _far_ too much when they just
need to know that a waiter is already running.

To avoid this big scary log message, we would need to set a log handler,
which would then need to either discard or redirect _all_ logs, or would
need to perform JSON parsing in order to deselect that particular
message. Parsing JSON in a log handler is very unappealing, I don't
really want this program logging to a file if it can be avoided, and
discarding all logs (some of which may help find legitimate bugs) is out
of the question. Therefore, we should do this the wrong way and try to
bind to the port ourselves.

Most of the time there won't be another waiter anyway, and we'll just
waste some time. If there is another waiter, then we write a clean
message for the user, and in the very unlikely event that there wasn't
another waiter but then one appears just before we actually start the
service, then we'll get the scary log message. In the also unlikely
event that the waiter ends just afterwards, then the clean message may
be a little confusing but presumably the user would just try again.
I think this behaviour would be fine, and preferable to consistently
getting a long error message that looks like there's a significant
issue, when in reality the user just already had the program running.

On Windows, the way to attempt to bind to a socket involves
Windows Sockets 2. These functions don't use the same machinery that
C++/WinRT assists with, so we're on our own, and including the
documented header `WinSock2.h` completely breaks the build anyway.
I don't really want to deal with this, so the error message is staying.

At least until I replace gRPC with something else, because the pain of
just getting it to do what I want [1] has overtaken the benefit of me
using it before.

[1] This includes not embedding the full path to each gRPC source file
    in the _release_ build, which I didn't think it was doing, but it
    turns out that the path just isn't in the JSON and is still
    delivered to the log handler.