Remove gRPC implementation, use Boost.Interprocess
As mentioned in fa44a009, gRPC is actually not as suitable as I had
hoped for this application. Instead of gRPC, use Boost.Interprocess to
manage a shared queue. Setters push messages to this queue when the
active profile changes, and waiters pull from the queue to be informed
of those changes. This makes a very simple RPC mechanism that does not
require any networking, works for only one RPC method, and does not
support return values. Protobuf is still used to serialize the message.