Reduce chance of queue not closing, allow recovery
Because the queue removal was not managed through RAII it would not
happen if the waiter exited by exception [1]. Use a `QueueHolder`
type to ensure that the sole waiter removes the queue when it exits,
including by an exception. `QueueHolder` uses the construction of the
queue to determine if the queue is 'owned' and therefore whether it
should be removed. Proper setters own the queue object but not the queue
itself and therefore do not remove it.

It is still possible for the queue to not be removed if the process
exits without calling destructors, such as by receiving SIGINT.
While we could try and detect this and clean up on next start, doing so
reliably is quite difficult and feels disproportionate for this program.
Instead, the queue not being removed will mean that no waiters can
start (proper setters will still function), and the waiter will prompt
the user to remove the queue if they know that the waiter did exit
unexpectedly. It's possible for the user to run two waiters by doing
this, but it should be hard to do accidentally.

[1] By this I mean not hitting the happy return path; `main` should
    catch uncaught exceptions so the process will not actually exit
    with an exception in the literal sense.